+++ src/dash-table/derived/edges/index.ts
@@ -3,38 +3,7 @@ import * as R from 'ramda';
 import { BorderStyle, BORDER_PROPERTIES } from './type';
 import { IConvertedStyle } from '../style';
 import { Datum, IVisibleColumn } from 'dash-table/components/Table/props';
-
-export type Filter<T> = (s: T[]) => T[];
-
-export const matchesDataCell = (datum: Datum, i: number, column: IVisibleColumn): Filter<IConvertedStyle> => R.filter<IConvertedStyle>((style =>
-    style.matchesRow(i) &&
-    style.matchesColumn(column) &&
-    style.matchesFilter(datum)
-));
-
-export const matchesFilterCell = (column: IVisibleColumn): Filter<IConvertedStyle> => R.filter<IConvertedStyle>((style =>
-    style.matchesColumn(column)
-));
-
-export const matchesHeaderCell = (i: number, column: IVisibleColumn): Filter<IConvertedStyle> => R.filter<IConvertedStyle>((style =>
-    style.matchesRow(i) &&
-    style.matchesColumn(column)
-));
-
-export const matchesDataOpCell = (datum: Datum, i: number): Filter<IConvertedStyle> => R.filter<IConvertedStyle>((style =>
-    !style.checksColumn() &&
-    style.matchesRow(i) &&
-    style.matchesFilter(datum)
-));
-
-export const getFilterOpStyles: Filter<IConvertedStyle> = R.filter<IConvertedStyle>((style =>
-    !style.checksColumn()
-));
-
-export const getHeaderOpStyles = (i: number): Filter<IConvertedStyle> => R.filter<IConvertedStyle>((style =>
-    style.matchesRow(i) &&
-    !style.checksColumn()
-));
+import { matchesDataCell, matchesDataOpCell, matchesFilterCell, getFilterOpStyles, matchesHeaderCell, getHeaderOpStyles } from 'dash-table/conditional';
 
 function resolveEdges(styles: IConvertedStyle[]): BorderStyle {
     let res: BorderStyle = {};
